# performance: (~3x) optimize field multiplier and EC-adder for improved MSM and ECNTT