--- a/icicle/backend/cpu/include/ntt_cpu.h
+++ b/icicle/backend/cpu/include/ntt_cpu.h
@@ -3,9 +3,16 @@
 #include "icicle/utils/log.h"
 #include "ntt_tasks_manager.h"
 #include "ntt_utils.h"
-// #include <_types/_uint32_t.h>
 #include <cstdint>
 #include <deque>
+
+#ifdef CURVE_ID
+  #include "icicle/curves/curve_config.h"
+using namespace curve_config;
+  #define IS_ECNTT std::is_same_v<E, projective_t>
+#else
+  #define IS_ECNTT false
+#endif
 
 using namespace field_config;
 using namespace icicle;
@@ -464,10 +471,15 @@ namespace ntt_cpu {
   bool NttCpu<S, E>::compute_if_is_parallel(uint32_t logn, const NTTConfig<S>& config)
   {
     uint32_t log_batch_size = uint32_t(log2(config.batch_size));
-    uint32_t scalar_size = sizeof(S);
-    // for small scalars, the threshold for when it is faster to use parallel NTT is higher
-    if ((scalar_size >= 32 && (logn + log_batch_size) <= 13) || (scalar_size < 32 && (logn + log_batch_size) <= 16)) {
-      return false;
+    // For ecntt we want parallelism unless really small case
+    if constexpr (IS_ECNTT) {
+      return logn > 5;
+    } else {
+      uint32_t scalar_size = sizeof(S);
+      // for small scalars, the threshold for when it is faster to use parallel NTT is higher
+      if ((scalar_size >= 32 && (logn + log_batch_size) <= 13) || (scalar_size < 32 && (logn + log_batch_size) <= 16)) {
+        return false;
+      }
     }
     return true;
   }