## Describe the changes

1. fix: ECNTT 5<logn<14 was single threaded. Now multi threaded.

2. This PR is using inlining and loop unrolling to improve CPU for:

- scalar multiplication
- scalar * ECpoint
- EC addition
- and consequently ECNTT and and MSM.


ECNTT (MULTI-THREAD) {size=2^14, curve=BN254}:

- For i9-13900k ubuntu22, clang-14: ~1200ms -> ~450ms (~2.6x)
- For m1-pro macos-15, clang-16: ~3500ms -> ~1200ms (~2.9x)


MSM (SINGLE-THREAD!) BLS12-377 logn=20:

- For i9-13900k ubuntu22, clang-14: ~32s -> ~10s (~3.2x)
- For m1-pro macos-15, clang-16:  ~36.5S  -> 13.3s (~2.7x)

MSM (MULTI-THREAD!) BLS12-377 logn=20:
- For i9-13900k ubuntu22, clang-14: ~1.9s -> ~1.9s (~1x) 
- For m1-pro macos-15, clang-16: ~5500ms -> ~3100ms (~1.8x)


CONCLUSION: ICICLE msm is bound by the manager thread. This is why for single thread we see 3X but for multi-thread mac (8-10 cores) improves 1.8x but i9 (32 cores) show no improvement. @LeonHibnik @mickeyasa @Koren-Brand